requirements.txt
aioredis==1.3.1
amqp==2.6.1
appdirs==1.4.4
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
autopep8==1.5.2
beautifulsoup4==4.9.1
billiard==3.6.3.0
bs4==0.0.1
cachetools==4.1.0
celery==4.4.7
certifi==2020.4.5
cffi==1.14.0
channels==2.4.0
channels-redis==3.0.1
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
daphne==2.5.0
decorator==4.4.2
Django==3.0.5
django-admin-numeric-filter==0.1.6
django-autocomplete-light==3.5.1
django-ckeditor==5.9.0
django-cors-headers==3.5.0
django-filter==2.2.0
django-floppyforms==1.8.0
django-haystack==2.8.1
django-js-asset==1.2.2
django-mathfilters==1.0.0
django-pagination-bootstrap==1.3
django-range-slider==0.2.4
django-treebeard==4.3.1
djangorestframework==3.12.2
dnspython==2.1.0rc1
fake-useragent==0.1.11
hiredis==1.1.0
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
kombu==4.6.11
lxml==4.6.2
msgpack==1.0.0
parse==1.18.0
Pillow==8.0.1
psycopg2==2.8.6
pycodestyle==2.5.0
pycparser==2.20
pyee==7.0.4
PyHamcrest==2.0.2
pyppeteer==0.2.2
pyquery==1.4.3
python-http-client==3.3.1
pytz==2019.3
requests==2.25.0
requests-html==0.10.0
sendgrid==6.4.8
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
starkbank-ecdsa==1.1.0
stripe==2.55.1
syllables==0.1.0
tqdm==4.54.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.26.2
vine==1.3.0
w3lib==1.22.0
websockets==8.1
zope.interface==5.1.0
